Move `initialize_tiled_datasets` to `tiled_datasets`
