+++ src/utils.py
@@ -17,7 +17,6 @@ from parameters import (
     TUNet3PlusParameters,
     TUNetParameters,
 )
-from tiled_dataset import TiledDataset
 
 
 def validate_parameters(parameters):
@@ -55,36 +54,6 @@ def validate_parameters(parameters):
 
     print("Parameters loaded successfully.")
     return io_parameters, network, model_parameters
-
-
-def initialize_tiled_datasets(
-    io_parameters, is_training=False, is_full_inference=False
-):
-    """
-    This function takes tiled uris from the io_parameter class, build the client and construct TiledDataset.
-    Input:
-        io_parameters: class, all io parameters in pydantic class
-        is_training: bool, whether the dataset is used for training or inference
-        is_full_inference: bool, whether the process is used for full inference
-    Output:
-        dataset: class, TiledDataset
-
-    """
-    data_tiled_client = from_uri(
-        io_parameters.data_tiled_uri, api_key=io_parameters.data_tiled_api_key
-    )
-    mask_tiled_client = None
-    if io_parameters.mask_tiled_uri:
-        mask_tiled_client = from_uri(
-            io_parameters.mask_tiled_uri, api_key=io_parameters.mask_tiled_api_key
-        )
-    dataset = TiledDataset(
-        data_tiled_client=data_tiled_client,
-        mask_tiled_client=mask_tiled_client,
-        is_training=is_training,
-        is_full_inference=is_full_inference,
-    )
-    return dataset
 
 
 def normalization(image):
